+++ frontend/components/admin/DeleteCategory/DeleteCategory.js
