+++ frontend/components/admin/DeleteMeat/DeleteMeat.js
