--- frontend/components/admin/meat/MeatGrid.js
@@ -1,105 +1,0 @@
-import React, { Component } from 'react';
-import { Query } from 'react-apollo';
-import { ALL_MEATS_QUERY } from '../../../queries/Meat';
-import { AdminGrid } from '../../styles/AdminGrid';
-import { DeleteMeat } from './DeleteMeat';
-import { EditMeat } from './EditMeat';
-import { ModalWindow } from '../../elements/ModalWindow';
-import { LoadingBox } from '../../elements/LoadingBox';
-
-class MeatGrid extends Component {
-    state = {
-        selected: {
-            id: '',
-            name: ''
-        }
-    };
-
-    // eslint-disable-next-line class-methods-use-this
-    showEditMeatForm(meat) {
-        document.getElementById('page-overlay').style.display = 'block';
-        document.getElementById('edit-meat-window').style.display = 'block';
-        document.getElementById('edit-meat-name').focus();
-
-        this.setState({ selected: { id: meat.id, name: meat.name } });
-    }
-
-    render() {
-        const gridStyle = {
-            width: '550px'
-        };
-        return (
-            <>
-                <ModalWindow id="edit-meat-window" width="500" height="215">
-                    <EditMeat id={this.state.selected.id} name={this.state.selected.name} />
-                </ModalWindow>
-
-                <AdminGrid cellPadding="0" cellSpacing="0" id="meatadmingrid" style={gridStyle}>
-                    <thead>
-                        <tr>
-                            <th width="64%" className="no-border">
-                                Name
-                            </th>
-                            <th width="18%" className="no-border">
-                                &nbsp;
-                            </th>
-                            <th width="18%">&nbsp;</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <Query query={ALL_MEATS_QUERY}>
-                            {({ data, error, loading }) => {
-                                if (loading)
-                                    return (
-                                        <tr>
-                                            <td colSpan="3">
-                                                <LoadingBox />
-                                            </td>
-                                        </tr>
-                                    );
-                                if (error)
-                                    return (
-                                        <tr>
-                                            <td colSpan="3">Error: {error.message}</td>
-                                        </tr>
-                                    );
-                                return data.meats.length > 0 ? (
-                                    data.meats.map(meat => (
-                                        <tr key={meat.id} id={meat.id}>
-                                            <td>{meat.name}</td>
-                                            <td align="center">
-                                                <button
-                                                    type="button"
-                                                    data-id={meat.id}
-                                                    onClick={e => {
-                                                        e.preventDefault();
-                                                        this.showEditMeatForm(meat);
-                                                    }}
-                                                >
-                                                    Edit
-                                                </button>
-                                            </td>
-                                            <td align="center">
-                                                <DeleteMeat id={meat.id} name={meat.name}>
-                                                    Delete
-                                                </DeleteMeat>
-                                            </td>
-                                        </tr>
-                                    ))
-                                ) : (
-                                    <tr>
-                                        <td colSpan="3">
-                                            <em>No Meats Defined</em>
-                                        </td>
-                                    </tr>
-                                );
-                            }}
-                        </Query>
-                    </tbody>
-                </AdminGrid>
-            </>
-        );
-    }
-}
-
-export { MeatGrid };
